Minor adjustments to WP Cron CLI-command class
